kstock_reporter/requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3
pykrx==1.0.47
gunicorn==21.2.0

# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.1.0
httpx==0.27.2

# Code Quality
flake8==7.1.1
black==24.10.0
mypy==1.13.0
django-stubs==5.1.1

# Monitoring
flower==2.0.1
sentry-sdk==2.18.0
django-prometheus==2.3.1

# Rate Limiting
slowapi==0.1.9
